Split tests into static / instance
